CORE-270 #sync dev and prod

+++ edu.harvard.i2b2.crc/src/server/edu/harvard/i2b2/crc/ejb/QueryInfoBean.java
@@ -246,19 +246,24 @@ public class QueryInfoBean { //implements SessionBean {
 	public MasterResponseType deleteQueryMaster(
 			DataSourceLookup dataSourceLookup, String userId, String masterId)
 			throws I2B2Exception {
+		if (masterId==null){
+			log.debug("Null master id sent to deleteQueryMaster method");
+			return null;
+		}
+		else {
+			SetFinderDAOFactory sfDaoFactory = this.getSetFinderDaoFactory(
+					dataSourceLookup.getDomainId(), dataSourceLookup
+					.getProjectPath(), dataSourceLookup.getOwnerId());
+			IQueryMasterDao queryMasterDao = sfDaoFactory.getQueryMasterDAO();
 
-		SetFinderDAOFactory sfDaoFactory = this.getSetFinderDaoFactory(
-				dataSourceLookup.getDomainId(), dataSourceLookup
-						.getProjectPath(), dataSourceLookup.getOwnerId());
-		IQueryMasterDao queryMasterDao = sfDaoFactory.getQueryMasterDAO();
+			queryMasterDao.deleteQuery(masterId);
 
-		queryMasterDao.deleteQuery(masterId);
-
-		MasterResponseType masterResponseType = new MasterResponseType();
-		QueryMasterType queryMasterType = new QueryMasterType();
-		queryMasterType.setQueryMasterId(masterId);
-		masterResponseType.getQueryMaster().add(queryMasterType);
-		return masterResponseType;
+			MasterResponseType masterResponseType = new MasterResponseType();
+			QueryMasterType queryMasterType = new QueryMasterType();
+			queryMasterType.setQueryMasterId(masterId);
+			masterResponseType.getQueryMaster().add(queryMasterType);
+			return masterResponseType;
+		}
 	}
 
 	/**
